Swiggy Task/src/com/swiggytask/Player.java
package com.swiggytask;

public class Player {
    private int health;
    private int strength;
    private int attack;

    public Player(int health, int strength, int attack) {
        this.health = health;
        this.strength = strength;
        this.attack = attack;
    }

    public int getHealth() {
        return health;
    }

    public void defend(int damage) {
        this.health -= damage;
    }

    public boolean isAlive() {
        return this.health > 0;
    }

    public int attack() {
        return attack * Dice.roll();
    }

    public int defend() {
        return strength * Dice.roll();
    }
}